+++ tests/docopt/test_advanced.py
@@ -1,5 +1,0 @@
-"""Tests an advanced parser
-"""
-from __future__ import annotations
-
-# TODO